API

Installer/Uninstaller Functions

TestTools.install — Function

TestTools.install(; <keyword arguments>)

Install all of the CLI utilities.

Keyword arguments

julia::AbstractString: path to julia executable. Default: path of the current running julia

bin_dir::AbstractString: directory to install CLI utilities into. Default: ~/.julia/bin

julia_flags::Vector{<:AbstractString}: command line flags for CLI executables. Default: ["--startup-file=no", "-q", "--compile=min", "-O0"]

force::Bool: flag used to indicate that existing CLI executables should be overwritten. Default: false

TestTools.uninstall — Function

TestTools.uninstall(; <keyword arguments>)

Unnstall all of the CLI utilities.

Keyword arguments

bin_dir::AbstractString: directory containing CLI executables to uninstall. Default: ~/.julia/bin

jltest

TestTools.jltest — Module

The jltest module provides support for running unit tests.

TestTools.jltest.find_tests — Function

find_tests(dir::AbstractString), <keyword arguments>)::Vector{String}

Construct a list of absolute paths to Julia test files contained in dir.

Keyword Arguments

recursive::Bool: flag indicating whether or not tests found in subdirectories of dir should be included in results. Default: true

exclude_runtests::Bool: flag indicating whether or not to exclude files named runtests.jl from the list of test files. Default: true

TestTools.jltest.run_tests — Function

run_tests(tests::Vector; <keyword arguments>)
run_tests(tests::AbstractString; <keyword arguments>)

Run tests in the list of files or modules provided in tests. If tests is an empty list or an empty string, an ArgumentError is thrown. File names in tests may be specified with or without the .jl extension.

Keyword Arguments

desc::AbstractString: description to use for test set used to group tests. Default: the default description set by @testset

test_set_type::Type: type of test set to use to group tests. When test_set_type is set to nothing, the tests are run individually. Default: EnhancedTestSet{DefaultTestSet}

Note

When the JLTEST_FAIL_FAST environment variable is set to "true", the test_set_type argument is overridden and set to EnhancedTestSet{Test.FallbackTestSet} so that tests are run in fail-fast mode.

recursive::Bool: flag indicating whether or not to run tests found in subdirectories of directories in tests. Default: true

exclude_runtests::Bool: flag indicating whether or not to exclude files named runtests.jl from the list of test files that are run. Default: true

jlcoverage

TestTools.jlcoverage — Module

The jlcoverage module provides support for analyzing code coverage of unit tests.

TestTools.jlcoverage.display_coverage — Function

display_coverage(coverage_data::Vector; startpath::AbstractString)

Display coverage results provided in coverage_data. File names are displayed relative to startpath. To display absolute paths, set startpath to an empty string. By default, startpath is set to the current working directory.

EnhancedTestSet

TestTools.jltest.EnhancedTestSet — Type

struct EnhancedTestSet{T<:AbstractTestSet} <: AbstractTestSet

Extension of the AbstracctTestSet type that provides the following functionality:

display diffs (when available) for comparison test failures;

support fail-fast (i.e., stop testing at first failure).

TestTools.jltest.EnhancedTestSet — Method

EnhancedTestSet(description::AbstractString; <keyword arguments>)

Construct an EnhancedTestSet with the specified description.

Keyword Arguments

wrap::Type{<:AbstractTestSet}: test set type to wrap. Default: DefaultTestSet