API

TestTools

TestTools.install — Function

TestTools.install(; <keyword arguments>)

Install all of the CLI utilities.

Keyword arguments

julia::AbstractString: path to julia executable. Default: path of the current running julia

bin_dir::AbstractString: directory to install CLI utilities into. Default: ~/.julia/bin

julia_flags::Vector{<:AbstractString}: command line flags for CLI executables. Default: ["--startup-file=no", "-q", "--compile=min", "-O0"]

force::Bool: flag used to indicate that existing CLI executables should be overwritten. Default: false

TestTools.uninstall — Function

TestTools.uninstall(; <keyword arguments>)

Unnstall all of the CLI utilities.

Keyword arguments

bin_dir::AbstractString: directory containing CLI executables to uninstall. Default: ~/.julia/bin

jltest

TestTools.jltest — Module

The jltest module provides support for running unit tests.

TestTools.jltest.TestSetPlus — Type

struct TestSetPlus{T<:AbstractTestSet} <: AbstractTestSet

Extension of the TestSet type that provides the following functionality:

display diffs when comparison tests fail and

supports "fail fast".

TestTools.jltest.find_tests — Method

find_tests(dir::AbstractString), <keyword arguments>)::Vector{String}

Recursively search dir for Julia files tests.

Keyword Arguments

exclude_runtests::Bool: flag that determines whether to exclude files named runtests.jl from the list of test files. Default: true

TestTools.jltest.run_tests — Method

run_tests(tests::Vector{<:AbstractString}; <keyword arguments>)
run_tests(tests::AbstractString; <keyword arguments>)

Run tests contained in the list of files or modules provided in tests. If tests is an empty list or an empty string, an ArgumentError is thrown. File names in tests may be specified with or without the .jl extension.

Keyword Arguments

name::AbstractString: name to use for test set used to group tests. Default: empty string

test_set_type::Type: type of test set to use to group tests. Default: TestSetPlus

jlcoverage

TestTools.jlcoverage — Module

The jlcoverage module provides support for analyzing code coverage of unit tests.

TestTools.jlcoverage.display_coverage — Function

display_coverage(coverage_data::Vector; startpath::AbstractString)

Display coverage results provided in coverage_data. File names are displayed relative to startpath. To display absolute paths, set startpath to an empty string. By default, startpath is set to the current working directory.

jlcodestyle

TestTools.jlcodestyle — Module

The jlcodestyle module provides support for checking and enforcing code style.